+++ ezgradle-plugin/src/main/java/me/thens/ezgradle/EzGradlePlugin.kt
@@ -1,5 +1,7 @@
 package me.thens.ezgradle
 
+import dev.horizona.core.log.LogFilter
+import me.thens.ezgradle.config.BuildConfigGenerator
 import me.thens.ezgradle.config.configureAndroidApplication
 import me.thens.ezgradle.config.configureAndroidLibrary
 import me.thens.ezgradle.config.configureHilt
@@ -9,31 +11,38 @@ import me.thens.ezgradle.config.configureKapt
 import me.thens.ezgradle.config.configureKotlin
 import me.thens.ezgradle.config.configureMavenPublish
 import me.thens.ezgradle.lib.EzGradleBomManager
-import me.thens.ezgradle.util.GenerateBuildConfigTask
+import me.thens.ezgradle.log.InternalLogger
+import me.thens.ezgradle.log.log
+import me.thens.ezgradle.model.EzGradleProperties
 import me.thens.ezgradle.util.isAndroid
 import me.thens.ezgradle.util.isJava
-import me.thens.ezgradle.model.EzGradleProperties
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 
 class EzGradlePlugin : Plugin<Project> {
     override fun apply(target: Project) {
         val ezgradleProperties = EzGradleProperties.from(target)
+        InternalLogger.update {
+            copy(printer = printer.filter(LogFilter.minLevel(ezgradleProperties.logLevel)))
+        }
+        log.debug { ezgradleProperties.toString() }
         target.configure(ezgradleProperties)
     }
 
-    private fun Project.configure(ezGradleProperties: EzGradleProperties) {
-        configureAndroidApplication()
-        configureAndroidLibrary()
+    private fun Project.configure(properties: EzGradleProperties) {
+        configureAndroidApplication(properties)
+        configureAndroidLibrary(properties)
         configureHilt()
         configureJava()
         configureJavaPlatform()
         configureKotlin()
         configureKapt()
         configureMavenPublish()
-        EzGradleBomManager(this).addDependencies(ezGradleProperties.bomVersion)
-        if (ezGradleProperties.generateBuildConfig && isJava && !isAndroid) {
-            GenerateBuildConfigTask.register(this)
+        if (properties.enableBom) {
+            EzGradleBomManager(this).addDependencies(properties.libs.versions.ezgradle)
+        }
+        if (properties.generateBuildConfig && isJava && !isAndroid) {
+            BuildConfigGenerator.registerTask(this)
         }
     }