+++ ezgradle-plugin/src/main/java/me/thens/ezgradle/log/Log.kt
@@ -1,0 +1,13 @@
+package dev.horizona.core.log
+
+enum class LogLevel { VERBOSE, DEBUG, INFO, WARN, ERROR, NEVER }
+
+typealias MessageProvider = () -> String?
+
+data class Log(
+    val level: LogLevel,
+    val tag: String,
+    val msg: MessageProvider,
+) {
+    fun fossil() = msg().let { copy(msg = { it }) }
+}