+++ ezgradle-plugin/src/main/java/me/thens/ezgradle/log/LogPrinterEmpty.kt
@@ -1,0 +1,6 @@
+package dev.horizona.core.log
+
+internal object LogPrinterEmpty : LogPrinter {
+    override fun print(log: Log) {
+    }
+}