+++ ezgradle-plugin/src/main/java/me/thens/ezgradle/model/ProjectProperties.kt
@@ -1,41 +1,46 @@
 package me.thens.ezgradle.model
 
+import me.thens.ezgradle.util.loadProperties
 import me.thens.ezgradle.util.toEnum
 import me.thens.ezgradle.util.toEnvName
 import org.gradle.api.Project
 
-class ProjectProperties(val project: Project, val prefix: String = "") {
+data class ProjectProperties(
+    val project: Project,
+    val properties: Map<String, String>,
+    val prefix: String,
+) {
     fun with(nextPrefix: String): ProjectProperties {
-        return ProjectProperties(project, "$prefix$nextPrefix")
+        return copy(prefix = "$prefix$nextPrefix")
     }
 
-    fun getString(name: String): String? {
+    operator fun get(name: String): String? {
         val key = "$prefix$name"
-        return project.findProject(key)?.toString()
-            ?: System.getenv(key.toEnvName())
+        val envName = key.toEnvName()
+        return System.getenv(envName) ?: listOf(key, envName)
+            .flatMap {
+                listOfNotNull(
+                    System.getProperty(it),
+                    properties[it],
+                    project.findProperty(it)?.toString(),
+                )
+            }
+            .firstOrNull()
     }
 
-    fun getBoolean(name: String): Boolean? {
-        return getString(name)?.toBoolean()
-    }
-
-    fun getInt(name: String): Int? {
-        return getString(name)?.toIntOrNull()
-    }
-
-    fun getLong(name: String): Long? {
-        return getString(name)?.toLongOrNull()
-    }
-
-    fun getFloat(name: String): Float? {
-        return getString(name)?.toFloatOrNull()
-    }
-
-    fun getDouble(name: String): Double? {
-        return getString(name)?.toDoubleOrNull()
+    fun getValue(name: String): String {
+        return get(name)
+            ?: throw NullPointerException("Cannot find value for key '$name in ProjectProperties")
     }
 
     fun <E : Enum<E>> getEnum(name: String, defaultValue: E): E {
-        return getString(name)?.toEnum(defaultValue.javaClass) ?: defaultValue
+        return get(name)?.toEnum(defaultValue.javaClass) ?: defaultValue
+    }
+
+    companion object {
+        fun from(project: Project): ProjectProperties {
+            val properties = project.loadProperties("local.properties")
+            return ProjectProperties(project, properties, "")
+        }
     }
 }