+++ ezgradle-run/src/main/java/me/thens/ezgradle/EzGradleTask.kt
@@ -1,0 +1,5 @@
+package me.thens.ezgradle
+
+fun interface EzGradleTask {
+    fun run(argv: List<String>)
+}