+++ ChangeLog.md
@@ -2,12 +2,22 @@
 
 ## TODO
 
-- Add `ezgradle.logLevel` to modify the log level
+- Migrate platform to version catalog
 - Add Git properties to BuildConfig class
 - Add `outdatedDependencies` task for project ezgradle-bom to show outdated dependencies
 - Add `updateDependencies` task for project ezgradle-bom to update outdated dependencies automatically
 
 ## 2024
+
+### 2024.03.01
+
+- Feature: Experimental for version catalog
+- Feature: Add `ezgradle.logLevel` to modify the log level
+- Refactor: Generate BuildConfig.java instead of BuildConfig.kt
+
+### 2024.02.02
+
+- Feature: Update supabase-bom from 1.4.7 to 2.2.2
 
 ### 2024.02.01